создает BoardController
